Moved file to right folder
Too difficult to use Firefox's in browser models, so I'm working on a special version that collects common words from a wikipedia list, then pinnging wiktionary for each word. Each wiktionary page has a bunch of tables with a bunch of translations into foreign languages. I take the first table as the best translation, but I hjold on to every translation i find for a special option (todo) of randomly choosing translations. This is the code to scrape that data and save new language files.