BIG update
Got mutations to work! pages will translate much better now, including on Twitter. However, there are still bugs with the quiz mode. If one switches into and out of quiz mode, and changes the vocab statistics etc., it will break for that page. Dunno why. But otherwise it works well for normal usage. 

Working on cleaning up the text of the popup.